Migrate ContentHeader to SCSS
